Move glossary related actions out of AdminController...
... into GlossaryController, and consolidate views.

Note two important loose ends:
1. The edit/update actions need to be secured with pundit.
2. The current glossary was hardcoded in /glossary/show.html.erb. Since
this commit changes that to be dynamic, we'll want to put the static
contents into a seed file or similar.
