--- app/views/admin/glossary/index.html.erb
@@ -1,4 +1,0 @@
-<%= render partial: 'admin/glossary/header' %>
-<%= link_to "Edit Glossary", glossary_admin_edit_path, class: "btn" %>
-<hr>
-<%= @glossary_content %>
