+++ app/views/glossary/edit.html.erb
@@ -1,8 +1,8 @@
-<%= render partial: 'admin/glossary/header' %>
-<%= link_to "View Glossary", glossary_admin_path %>
+<%= render partial: 'glossary/header' %>
+<%= link_to "View Glossary", glossary_path %>
 <hr>
 <div class="glossary_form">
-  <%= form_with model: @system_settings, url: glossary_admin_edit_path, local: true do |f| %>
+  <%= form_with model: @system_settings, url: glossary_path, local: true do |f| %>
     <div class="field">
       <%= f.rich_text_area :glossary_content %>
     </div>
