+++ app/views/glossary/show.html.erb
@@ -1,93 +1,4 @@
-<div class="columns">
-  <div class="column is-12">
-    <div class="title">Glossary</div>
-    <hr>
-  </div>
-</div>
-
-<div class="row columns">
-  <div class="columns column is-12">
-    <div class="column is-12">
-      <%= link_to "Contributions", contributions_path %>: All public-generated content, like, Asks, Offers, Community Resources, Announcements
-
-      <hr>
-
-      <%= link_to "Community Resources", community_resources_path(is_approved: false) %>: Local offerings and organizations that are always available, e.g. Food Pantry, Diaper Bank. Can be links or videos
-      <br>
-      <%= link_to "Announcements", announcements_path(is_approved: false) %>: References and notes to share with the public about specific events and offerings, e.g. food drive Jun 12
-
-      <hr>
-
-
-      <%= link_to "Asks", listings_path(contribution_type: "Ask") %>: Request for support/exchange from the public
-      <br>
-      <%= link_to "Offers", listings_path(contribution_type: "Offer") %> Offer of support/exchange from the public
-      <br>
-      <%= link_to "Donations", donations_path %>: Log of Donations to the mutual aid group, but this form could connect to Stripe, etc, later
-
-      <hr>
-
-      <%= link_to "Locations", locations_path %>: Address-related info submitted by People or Organizations when they fill out Forms. Also used to store Service Area info
-      <br>
-      <%= link_to "Location Types", location_types_path %>: Kind of Location, so, home, business, cross street, etc
-      <br>
-      <%= link_to "Submissions", submissions_path %>: Form "responses" (This is a log of all results from all Forms: Ask, Offer, (CommunityResource, Annoucement, Volunteer), etc)
-      <br>
-      <%= link_to "Form Questions", custom_form_questions_path %>: Questions you want to be on your Forms
-      <br>
-      <%= link_to "Submission Responses", submission_responses_path %>: People's answers to specific Form Questions
-
-      <hr>
-
-
-      <%= link_to "Matches", matches_path %>: Connection between a receiver (Ask/Profile) and a provider (Offer/Community Resource/Profile) (Profile matches are TBD)
-      <br>
-      <%= link_to "Feedback", feedbacks_path %>: Each Match will generate an email request to receivers and providers to request Feedback about their experience. (We are considering showing stats on Profiles re thumbsup/starts/rating, TBD)
-      <br>
-      <%= link_to "Communication Log", communication_logs_path %>: Log of all phone calls, texts, etc, to People. Any system-generated emails will be saved as such.
-
-      <hr>
-
-      <%= link_to "People", people_path %>: All the human People in the system -- receiver/provider/leadership team member/partner organization point of contact/etc
-      <br>
-      <%= link_to "Volunteers", volunteer_admin_path %>: People who are part of the group (different than Providers who submit Offers)
-      <br>
-      <%= link_to "Positions", positions_path %>: Title/Role/Membership a Person has within an Organization, e.g. Leadership Team member/Dispatch Team member/Point of Contact
-
-      <hr>
-
-      <%= link_to "User accounts", users_path %>: For People who log in to this system. People will still have Profiles even if they don't have a User Account. (Still need to decide depending on login status.)
-      <br>
-      <%= link_to "Shared Accounts", shared_accounts_path %>: List of accounts the mutual aid group keeps, along w descriptions (but NOT passwords!), so there's one list of all the gmail addresses, payment gateways, etc. (Again, NOT the actual passwords!)
-      <br>
-      <%= link_to "Software Feedback", software_feedbacks_path %>: Place for Users (Admins only?) to store ideas/feedback for this system. (Y'all can decide which of these you want to request of coders...)
-      <br>
-
-      <%= link_to "Categories", categories_path %>: Services offered/facilitated by the mutual aid group. Can be nested. Categories will appear on Form checkboxes and dropdowns, and Contribution page filters
-      <br>
-      <%= link_to "Service Areas", service_areas_path %>: Geographic "areas" served by the mutual aid group. Can be nested. Can start at state, and drill down to Pods. Service Areas will appear on Form checkboxes and dropdowns, and Contribution page filters
-      <br>
-      <%= link_to "Contact Methods", contact_methods_path %>: Mechanism for contacting People, e.g. Text, Call, Email, WhatsApp. Contact Methods will appear on Form checkboxes and dropdowns, and on Contribution page filters
-
-      <hr>
-      <%= link_to "Translations", mobility_string_translations_path %>: Working up to multilanguage support
-      <br>
-      <%= link_to "System Locales", system_locales_path %>: Languages your group supports. We populated this list, but you can change it to whatever you want!
-
-      <hr>
-
-      <%= link_to "Yearbook", yearbook_path %>: List of everyone who has a Position. (Later we can have Profile photos.)
-
-      <hr>
-
-      <%= link_to "Respond", respond_contribution_path(Listing.first) %>: Preview a Contribution and prepare to "Respond" to the person by adding a Tentative Match
-      <br>
-      <%= link_to "Triage", triage_contribution_path(Listing.first) %>: Update Contribution -- Add short summary, change description, indicate if 'Inexhaustible'
-      <br>
-      <span class="has-text-weight-bold">Tentative Match:</span> This is a real Match. The first point of contact is to choose a potential match, and then to contact both parties.
-      <br>
-      <span class="has-text-weight-bold">Inexhaustible:</span> This Contribution can be matched many times, will continue to show up in dropdowns, etc
-
-    </div>
-  </div>
-</div>
+<%= render partial: 'glossary/header' %>
+<%= link_to "Edit Glossary", edit_glossary_path, class: "btn" %>
+<hr>
+<%= @system_settings.glossary_content %>
